action/requirements.txt
langchain>=0.2.1
langchain-openai>=0.2.1
openai>=1.1.0
PyGithub>=2.1.0